Split API with simulation engine
EngineAPI is now a default API that does nothing (but still useful
to be able to load and parse MORSE code outside Blender), and
BlenderAPI is the API to BlenderGameEngine.

The correct one is selected in morse.core.__init__.py

This commit is not complete. Several references to the
previous morse.core.blenderapi module must be updated.